Simulia: Version 1.2.0 Somes added features and an improvement of the precision : food is now a double and the food production is now more precise. New reaction system in case of shortage of food, the citizens will adapt and try to stop it.

+++ Resources/Naturals Resources.cs
@@ -1,6 +1,6 @@
 public class NaturalsResources
 {
-    public int Food { get; set; } = 10;
+    public double Food { get; set; } = 10;
     public int Water { get; set; } = 10;
     public int Wood { get; set; } = 10;
     public int Minerals { get; set; } = 10;
